Refactor entity properties to use long instead of short for IdPlatform, IdMembershipType, IdProductType, IdFormat, and IdAccountType. Update DatabaseContext and configuration classes for consistent naming conventions. Enhance Program.cs with service registrations and middleware for exception handling.

+++ Domain/Entities/GamePlatformEntity.cs
@@ -4,7 +4,7 @@ namespace VideoGameStore.Domain.Entities
     {
         public long IdGame { get; set; }
 
-        public short IdPlatform { get; set; }
+        public long IdPlatform { get; set; }
         
         public GameEntity? Game { get; set; }
