+++ Domain/Entities/PlatformEntity.cs
@@ -2,7 +2,7 @@ namespace VideoGameStore.Domain.Entities
 {
     public class PlatformEntity : BaseEntity
     {
-        public short IdPlatformOwner { get; set; }
+        public long IdPlatformOwner { get; set; }
 
         public string Name { get; set; } = string.Empty;
