+++ Domain/Entities/ProductVariantEntity.cs
@@ -4,7 +4,7 @@ namespace VideoGameStore.Domain.Entities
     {
         public long IdProduct { get; set; }
 
-        public short IdAccountType { get; set; }
+        public long IdAccountType { get; set; }
 
         public decimal Price { get; set; }
